Fix support fresh saves with HoF uninit sectors
Before entering the HoF for the first time, the related sectors are
untouched by the game and hence remain unitialized from NAND memory
which sets it to 0xFF.

Save files that have gone through PKHeX have still the HoF sectors all
set to 0xFF, except for the checksum field which is computed as 0xFC1F
from the 0xFF data portion. This means that they will also have the
security field set to 0xFFFFFFFF (invalid).

After entering the HoF for the first time, the second sector will be
semantically empty, but will have the data portion set to 0x00, valid
checksum (incidentally?) 0x00 and valid security field.